wrote base setup for parry & blocking states, added healthbar functionality

+++ Fighter/Assets/Scripts/BothChars/IHealth.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public interface IHealth{
 
-    void ChangeHealth(int damage);
+    void ChangeHealth(float damage);
     void OnDeath();
 
 }
